ISSUE #840 Support contesting sensor ownership (deadline 6th Dec 2022) Clean up and refactoring
