+++ app/src/main/java/com/ruuvi/station/network/di/NetworkInjectionModule.kt
@@ -11,7 +11,7 @@ object NetworkInjectionModule {
     val module = Kodein.Module(NetworkInjectionModule.javaClass.name) {
         bind<RuuviNetworkRepository>() with singleton { RuuviNetworkRepository(Dispatchers.IO, instance()) }
 
-        bind<SensorClaimInteractor>() with singleton { SensorClaimInteractor(instance(), instance(), instance(), instance()) }
+        bind<SensorClaimInteractor>() with singleton { SensorClaimInteractor(instance(), instance(), instance(), instance(), instance(), instance()) }
 
         bind<NetworkTokenRepository>() with singleton { NetworkTokenRepository(instance(), instance(), instance()) }
 
@@ -46,7 +46,7 @@ object NetworkInjectionModule {
         }
 
         bind<ClaimSensorViewModel>() with factory { sensorId: String ->
-            ClaimSensorViewModel(sensorId, instance(), instance(), instance(), instance(), instance(), instance())
+            ClaimSensorViewModel(sensorId, instance(), instance(), instance(), instance())
         }
 
         bind<NetworkResponseLocalizer>() with provider { NetworkResponseLocalizer(instance()) }
